[GCIPJ] Commit of Binary to 7 segments converter
This is a converter of a binary number to a signal that show the decimal number representation on a 7 segments display.

It is divided in two module. The first one, BIN_2_BCD, converts binary to BCD. The second one, BIN_2_7SEG uses the first to do the hole thing, convert binary into data that can be used by the 7 segments' displays.